revert latest change and added APPLICATION_LOGIN_LEFT_PANEL_SHOWN to configure the left panel visibilty
