+++ org.idempiere.zk.iceblue_c.theme/src/web/theme/iceblue_c/zul/login/login-links.zul
@@ -15,7 +15,8 @@ Copyright (C) 2009 Idalica Corporation
 	}
 </style>
 
-<div class="loginLinksBox" if="${desktop.execution.serverName.equals(&quot;127.0.0.1&quot;) or desktop.execution.serverName.startsWith(&quot;demo&quot;) or desktop.execution.serverName.startsWith(&quot;test.&quot;) or desktop.execution.serverName.startsWith(&quot;test-&quot;) or desktop.execution.serverName.startsWith(&quot;triage&quot;)}">
+<div class="loginLinksBox" if="${desktop.execution.serverName.equals(&quot;127.0.0.1&quot;) or desktop.execution.serverName.startsWith(&quot;demo&quot;) or desktop.execution.serverName.startsWith(&quot;test.&quot;) or desktop.execution.serverName.startsWith(&quot;test-&quot;) or desktop.execution.serverName.startsWith(&quot;triage&quot;)}"
+	style="position: absolute; bottom: 30px;">
 
 	<toolbarbutton href="http://www.zkoss.org" image="~./theme/iceblue_c/images/zkpowered_s.png"
 		tooltiptext="Powered by Zk" target="_blank"
